pyproject.toml
[tool.poetry]
name = "etl-ai-pipeline"
version = "0.1.0"
description = ""
authors = ["JakubKub13 <jakubkubala3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.38"
aiosqlite = "^0.21.0"
greenlet = "^3.1.1"
supabase = "^2.13.0"
tenacity = "^9.0.0"
langgraph = "^0.2.74"
langchain-anthropic = "^0.3.7"

[tool.poetry.scripts]
start = "etl_ai_pipeline.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
